recsys_python/requirements.txt
Flask
pymongo[srv]  
scikit-learn
pandas
numpy
scikit-surprise